Added not functual rotation and polygon filling that has to be done without the graphics class later.

+++ PseudoEngine/core/Camera.cs
@@ -22,7 +22,7 @@ namespace PseudoEngine.core
 
         #endregion
 
-        public Camera(int screenWidth, int screenHeight)
+        public Camera(double screenWidth, double screenHeight)
         {
             Ratio = screenWidth / screenHeight;
         }
